requirements.txt
python-telegram-bot==21.11.1
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.8.1
starlette==0.46.1
cryptography==44.0.2
asyncpg==0.30.0
